:recycle: rename login component to authentication
