+++ src/authentication/jwt.strategy.ts
