+++ src/users/user.decorator.ts
@@ -1,6 +1,6 @@
 import { createParamDecorator } from '@nestjs/common';
 import { Request } from 'express';
-import { UserBody } from 'src/login/dto/userbody.dto';
+import { UserBody } from 'src/authentication/dto/userbody.dto';
 
 export const LoggedUser = createParamDecorator(
   (data: unknown, req: Request): UserBody => {
